fix the lengths of weight tensor
